Find nearest point to lines

This will work if the lines are specified as a sp::SpatialLines or sp::SpatialPolygon object. If the latter, the object is converted to a sp::SpatialLines object. It will also work if the lines are specified in a dataframe (or matrix) with x, y, and id (optional), which is a coordinate matrix. The points can be a sp::SpatialPoints object, a dataframe or matrix with each row the x-y coordinates with an optional id for multiple lines or it can be a vector if it is only one point.

getNearestPointOnLine(
  p,
  l = buffer300.wintri$df,
  crs.to.use = buffer300.wintri$crs
)

Arguments

p
the points as a sp::SpatialPoints object, dataframe, matrix or vector (one point). If a dataframe or matrix, column 1 is the x locations, column 2 is the y locations. The units and projection of the points need to match the lines. If sp::SpatialPoints is used, this can be checked. Otherwise the user needs to ensure that they match.

l
the lines as a sp::SpatialPolygons object, sp::SpatialLines object, dataframe, matrix or vector (one point). If a dataframe or matrix, column 1 is the x locations, column 2 is the y locations. If there are multiple lines, the third column should be the id of the lines.

crs.to.use
name of the projection. must be one that the sp package recognizes.

Value

vector of points

Details

Note, the maptools::snapPointsToLines() function should do this, but it is not working with the SpatialLines object produced by converting a polygon to lines. It may work fine if one converts the SpatialPolygon to a SpatialLinesDataFrame with the proper line id's.

The default behavior is to use the 300km coastline buffer with the Winkel Tripel projection.