Get the mean raster values along a line

This is applies raster::extract() but does some checking on the projection of the inputs. Set d=0 to get value at the points.

getStatsAlongLine(r, p, l = NULL, d = 0, na.rm = TRUE)

Arguments

r
the raster as a raster::raster object

p
the points as a sp::SpatialPoints object

l
the lines as a sp::SpatialLines object if d is not 0.

d
distance to average along the line. Distance is in the units of the points and raster objects

na.rm
whether to remove NAs in the raster when computing the average

Value

raster values around the points

Details

If there are multiple lines within the radius of d, then the average is over all lines covered in that radius. The segments to average is determine by a circle around the points.

See also